Base code with pre-trained NN
The code as is uses an NN with 4 inputs: x and y of current robot position, and previous left and right motor speed. The 'best.er.20231018162343.com.nmu.evos.TanHActivation.com.nmu.evos.TanHActivation.nn' NN was trained for moving from the start to end point without obstacles in between. It can be tested in the 'Main' program.